Bitbucket

Documentation for Bitbucket.

Bitbucket.Person

Bitbucket.PullRequest

Bitbucket.User

Bitbucket.Authenticate

Bitbucket.GetRepoPRs

Bitbucket.GetRole

Bitbucket.GetRole

Bitbucket.GetUserPRs

Bitbucket.Person — Type

Fields

name: A username as it appears in BitBucket

email: The person's work email address

display_name: The person's full name

Bitbucket.PullRequest — Type

A struct to represent a single pull request.

Fields

link: The URL that points to the PR

Bitbucket.User — Type

A struct to represent the person running the queries.

Fields

username: A username as it appears in BitBucket

token: An HTTP token or the user's password. Using a token is recommended as you can control the permission level.

Bitbucket.Authenticate — Method

Authenticate(user::User)::AuthenticatedUser

Generates the auth token required by the API. Returns the user + token in a wrapper struct.

Bitbucket.GetRepoPRs — Method

GetRepoPRs(user::AuthenticatedUser, base_address::AbstractString, repo::Repository; state::AbstractString="ALL", start_date::Union{Date, DateTime}=(now()-Week(4)), end_date::Union{Date, DateTime}=now(), get_all::Bool=false, page_size::UInt=UInt(50))::Vector{PullRequest}

Get the pull requests for the given repository. Can specify optional arguments for filtering. By default will only return the last 4 weeks of data.

See also GetUserPRs

Bitbucket.GetRole — Method

GetRole(username::String, pr::PullRequest)::ParticipationType

Returns the person's role on the pull request. Returns NotOnPR if the person does not appear on the pull request.

Bitbucket.GetRole — Method

GetRole(username::String, pr::PullRequest)::ParticipationType

Returns the role of the person with the given username on the given pull request. Returns NotOnPR if the person does not appear on the pull request.

Bitbucket.GetUserPRs — Method

GetUserPRs(user::AuthenticatedUser, base_address::AbstractString; state::AbstractString="ALL", start_date::Union{Date, DateTime}=(now()-Week(4)), end_date::Union{Date, DateTime}=now(), get_all::Bool=false, page_size::UInt=UInt(50))::Vector{PullRequest}

Get the pull requests for the given user. Can specify optional arguments for filtering. By default will only return the last 4 weeks of data.

See also GetRepoPRs